requirements.txt
discord.py>=2.3.2,<3.0.0
python-dotenv>=1.0.1,<2.0.0
google-generativeai>=0.8.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
psutil>=5.9.0
aiofiles>=23.0.0
aiohttp>=3.9.0
